Moved tests to correct file Refactored tests
