+++ tests/cli/Cli/Console/RegisterModulesCest.php
@@ -29,7 +29,7 @@ class RegisterModulesCest
      * @author Nathan Edwards <https://github.com/npfedwards>
      * @since  2018-26-12
      */
-    public function cliConsoleRegisterModules(CliTester $I)
+    public function registerModules(CliTester $I)
     {
         $I->wantToTest("Cli\Console - registerModules()");
 
@@ -105,7 +105,7 @@ class RegisterModulesCest
      * @author Sid Roberts <https://github.com/SidRoberts>
      * @since  2019-05-15
      */
-    public function cliConsoleRegisterModulesBadPathThrowsAnException(CliTester $I)
+    public function registerModulesBadPathThrowsAnException(CliTester $I)
     {
         $I->wantToTest("Cli\Console - registerModules() - bad path throws exception");
